+++ app/controllers/admin/transfer_requests_controller.rb
@@ -527,7 +527,7 @@ module Admin
       end
 
       TransferRequestMailer.pending_lv_notification(tr).deliver_later
-      TransferRequestMailer.clubs_informed_lv_pending(tr).deliver_later
+      TransferRequestMailer.deliver_to_all_audiences(:clubs_informed_lv_pending, tr)
 
       redirect_to "#{base_url}?result=approved", allow_other_host: true
     end